+++ packages/components/src/placeholder/stories/index.story.tsx
@@ -45,6 +45,7 @@ const Template: StoryFn< typeof Placeholder > = ( args ) => {
 			<div>
 				<TextControl
 					__nextHasNoMarginBottom
+					__next40pxDefaultSize
 					label="Sample Field"
 					placeholder="Enter something here"
 					value={ value }