+++ packages/components/src/text-control/stories/index.story.tsx
@@ -54,6 +54,7 @@ export const Default: StoryFn< typeof TextControl > = DefaultTemplate.bind(
 );
 Default.args = {
 	__nextHasNoMarginBottom: true,
+	__next40pxDefaultSize: true,
 };
 
 export const WithLabelAndHelpText: StoryFn< typeof TextControl > =