+++ packages/components/src/text-control/test/text-control.tsx
@@ -9,7 +9,13 @@ import { render, screen } from '@testing-library/react';
 import _TextControl from '..';
 
 const TextControl = ( props: React.ComponentProps< typeof _TextControl > ) => {
-	return <_TextControl { ...props } __nextHasNoMarginBottom />;
+	return (
+		<_TextControl
+			{ ...props }
+			__nextHasNoMarginBottom
+			__next40pxDefaultSize
+		/>
+	);
 };
 
 const noop = () => {};